chore: Rename headlines to headings internally
