+++ components/o3-editorial-typography/heading.css
