+++ components/o3-editorial-typography/main.css
@@ -1,6 +1,6 @@
 @import './theme.css';
 @import './body.css';
-@import './headline.css';
+@import './heading.css';
 @import './details.css';
 @import './list.css';
 @import './link.css';
